計算が狂うのでツールバー文字列サイズを整数のみに限定するよう変更

+++ Pe-WPF/Pe/PeMain/Constants.cs
@@ -158,7 +158,7 @@
 		);
 
 		[ConstantsRange]
-		public static readonly TripleRange<double> toolbarTextLength = new TripleRange<double>(
+		public static readonly TripleRange<int> toolbarTextLength = new TripleRange<int>(
 			20,
 			80,
 			200
